+++ src/ObjectQR.ts
@@ -25,6 +25,8 @@ import {
     QRCodeInterface,
     QRCodeType,
     QRCodeSettings,
+    QRCodeDataSchema,
+    ExportObjectDataSchema
 } from '../index';
 
 export class ObjectQR extends QRCode implements QRCodeInterface {
@@ -55,21 +57,44 @@ export class ObjectQR extends QRCode implements QRCodeInterface {
     }
 
     /**
-     * The `toJSON()` method should return the JSON
-     * representation of the QR Code content.
+     * Parse a JSON QR code content into a ObjectQR
+     * object.
      *
-     * @return {string}
+     * @param   json        {string}
+     * @return  {ObjectQR}
+     * @throws  {Error}     On empty `json` given.
+     * @throws  {Error}     On missing `type` field value.
+     * @throws  {Error}     On unrecognized QR code `type` field value.
      */
-    public toJSON(): string {
+    static fromJSON(
+        json: string
+    ): ObjectQR {
 
-        const jsonSchema = {
-            'v': 3,
-            'type': this.type,
-            'network_id': this.networkType,
-            'chain_id': this.chainId,
-            'data': this.object,
-        };
+        // create the QRCode object from JSON
+        return ExportObjectDataSchema.parse(json);
+    }
 
-        return JSON.stringify(jsonSchema);
+    /**
+     * The `getTypeNumber()` method should return the
+     * version number for QR codes of the underlying class.
+     *
+     * @see https://en.wikipedia.org/wiki/QR_code#Storage
+     * @return {number}
+     */
+    public getTypeNumber(): number {
+        // Type version for ContactQR is Version 10
+        // This type of QR can hold up to 174 bytes of data.
+        return 10;
+    }
+
+    /**
+     * The `getSchema()` method should return an instance
+     * of a sub-class of QRCodeDataSchema which describes
+     * the QR Code data.
+     *
+     * @return {QRCodeDataSchema}
+     */
+    public getSchema(): QRCodeDataSchema {
+        return new ExportObjectDataSchema();
     }
 }