+++ src/QRCodeSettings.ts
@@ -36,21 +36,6 @@ export class QRCodeSettings {
     public static CORRECTION_LEVEL = ErrorCorrectLevel.L;
 
     /**
-     * The QR Code version number.
-     * 
-     * With `40-L` configuration, the QR Code can contain
-     * up to `2953` bytes. As defined in the following link,
-     * this is the maximum storage capacity for our types
-     * or QR Codes:
-     * 
-     *     https://en.wikipedia.org/wiki/QR_code#Design
-     *
-     * @see https://en.wikipedia.org/wiki/QR_code#Design
-     * @var {string}
-     */
-    public static VERSION_NUMBER = 40;
-
-    /**
      * The QR Code cell size in pixels.
      *
      * @var {number}