+++ test/ContactQR.spec.ts
@@ -18,18 +18,9 @@ import {
     PublicAccount,
     NetworkType,
 } from 'nem2-sdk';
-import {
-    QRCode as QRCodeImpl,
-    QR8BitByte,
-    ErrorCorrectLevel,
-} from 'qrcode-generator-ts';
 
 // internal dependencies
-import { 
-    QRCodeInterface,
-    QRCode,
-    QRCodeType,
-    QRCodeSettings,
+import {
     ContactQR,
 } from "../index";
 
@@ -39,13 +30,14 @@ describe('ContactQR -->', () => {
 
         it('include mandatory NIP-7 QR Code base fields', () => {
             // Arrange:
+            const name = 'test-contact-1';
             const account = PublicAccount.createFromPublicKey(
                 'C5C55181284607954E56CD46DE85F4F3EF4CC713CC2B95000FA741998558D268',
                 NetworkType.TEST_NET
             );
 
             // Act:
-            const addContact = new ContactQR(account, NetworkType.TEST_NET, '');
+            const addContact = new ContactQR(name, account, NetworkType.TEST_NET, '');
             const actualJSON = addContact.toJSON();
             const actualObject = JSON.parse(actualJSON);
 
@@ -55,6 +47,24 @@ describe('ContactQR -->', () => {
             expect(actualObject).to.have.property('network_id');
             expect(actualObject).to.have.property('chain_id');
             expect(actualObject).to.have.property('data');
+        });
+
+        it('include specialized schema fields', () => {
+            // Arrange:
+            const name = 'test-contact-1';
+            const account = PublicAccount.createFromPublicKey(
+                'C5C55181284607954E56CD46DE85F4F3EF4CC713CC2B95000FA741998558D268',
+                NetworkType.TEST_NET
+            );
+
+            // Act:
+            const addContact = new ContactQR(name, account, NetworkType.TEST_NET, '');
+            const actualJSON = addContact.toJSON();
+            const actualObject = JSON.parse(actualJSON);
+
+            // Assert:
+            expect(actualObject.data).to.have.property('name');
+            expect(actualObject.data).to.have.property('publicKey');
         });
     });