+++ test/CosignatureQR.spec.ts
@@ -15,6 +15,7 @@
  */
 import {expect} from "chai";
 import {
+    AggregateTransaction,
     TransferTransaction,
     Deadline,
     Address,
@@ -23,19 +24,29 @@ import {
     UInt64,
     PlainMessage,
     NetworkType,
+    PublicAccount,
 } from 'nem2-sdk';
+import {
+    QRCode as QRCodeImpl,
+    QR8BitByte,
+    ErrorCorrectLevel,
+} from 'qrcode-generator-ts';
 
 // internal dependencies
 import {
-    TransactionQR,
+    CosignatureQR,
 } from "../index";
 
-describe('TransactionQR -->', () => {
+describe('CosignatureQR -->', () => {
 
     describe('toJSON() should', () => {
 
         it('include mandatory NIP-7 QR Code base fields', () => {
             // Arrange:
+            const account = PublicAccount.createFromPublicKey(
+                'C5C55181284607954E56CD46DE85F4F3EF4CC713CC2B95000FA741998558D268',
+                NetworkType.MIJIN_TEST
+            );
             const transfer = TransferTransaction.create(
                 Deadline.create(),
                 Address.createFromPublicKey(
@@ -46,9 +57,14 @@ describe('TransactionQR -->', () => {
                 PlainMessage.create('Welcome to NEM!'),
                 NetworkType.MIJIN_TEST
             );
+            const bonded = AggregateTransaction.createBonded(
+                Deadline.create(),
+                [transfer.toAggregate(account)],
+                NetworkType.MIJIN_TEST
+            );
 
             // Act:
-            const requestTx = new TransactionQR(transfer, NetworkType.TEST_NET, '');
+            const requestTx = new CosignatureQR(bonded, NetworkType.TEST_NET, '');
             const actualJSON = requestTx.toJSON();
             const actualObject = JSON.parse(actualJSON);
 
@@ -62,6 +78,10 @@ describe('TransactionQR -->', () => {
 
         it('include specialized schema fields', () => {
             // Arrange:
+            const account = PublicAccount.createFromPublicKey(
+                'C5C55181284607954E56CD46DE85F4F3EF4CC713CC2B95000FA741998558D268',
+                NetworkType.MIJIN_TEST
+            );
             const transfer = TransferTransaction.create(
                 Deadline.create(),
                 Address.createFromPublicKey(
@@ -72,9 +92,14 @@ describe('TransactionQR -->', () => {
                 PlainMessage.create('Welcome to NEM!'),
                 NetworkType.MIJIN_TEST
             );
+            const bonded = AggregateTransaction.createBonded(
+                Deadline.create(),
+                [transfer.toAggregate(account)],
+                NetworkType.MIJIN_TEST
+            );
 
             // Act:
-            const requestTx = new TransactionQR(transfer, NetworkType.TEST_NET, '');
+            const requestTx = new CosignatureQR(bonded, NetworkType.TEST_NET, '');
             const actualJSON = requestTx.toJSON();
             const actualObject = JSON.parse(actualJSON);